requirements.txt
selenium==4.24.0
webdriver-manager==4.0.2
pytest==8.3.2
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
python-dotenv==1.0.1
allure-pytest==2.13.5
black==24.8.0
ruff==0.6.8